fix(scanner): address reference scan review [IDE-2493]
Clarify the reference-scan invariant and use durable temporary-folder terminology in tests. Move the folder scan context helper to focused functional options while preserving all scanner safeguards.
