+++ infrastructure/iac/reference_scan_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package oss
+package iac
 
 import (
 	"context"
@@ -31,7 +31,6 @@ import (
 	"github.com/snyk/snyk-ls/infrastructure/cli/mock_cli"
 	"github.com/snyk/snyk-ls/infrastructure/utils"
 	ctx2 "github.com/snyk/snyk-ls/internal/context"
-	"github.com/snyk/snyk-ls/internal/notification"
 	"github.com/snyk/snyk-ls/internal/observability/error_reporting"
 	"github.com/snyk/snyk-ls/internal/observability/performance"
 	"github.com/snyk/snyk-ls/internal/product"
@@ -39,24 +38,25 @@ import (
 	"github.com/snyk/snyk-ls/internal/types"
 )
 
-func newReferenceOSSScanner(
+func newReferenceIaCScanner(
 	t *testing.T,
 	token string,
-) (*CLIScanner, *bool, *types.ConfigResolver, types.FilePath) {
+) (*Scanner, *bool, *types.ConfigResolver, types.FilePath, types.FilePath) {
 	t.Helper()
 	engine, tokenService := testutil.UnitTestWithEngine(t)
 	tokenService.SetToken(engine.GetConfiguration(), token)
 	conf := engine.GetConfiguration()
-	conf.Set(configresolver.UserGlobalKey(types.SettingSnykOssEnabled), false)
+	conf.Set(configresolver.UserGlobalKey(types.SettingSnykIacEnabled), false)
 
 	realFolder := types.FilePath(t.TempDir())
 	referenceFolder := types.FilePath(t.TempDir())
-	require.NoError(t, os.WriteFile(filepath.Join(string(referenceFolder), "package.json"), []byte(`{"name":"reference"}`), 0o600))
-	types.SetFolderUserSetting(conf, realFolder, types.SettingSnykOssEnabled, true)
+	supportedPath := types.FilePath(filepath.Join(string(referenceFolder), "main.yml"))
+	require.NoError(t, os.WriteFile(string(supportedPath), []byte("resources: {}"), 0o600))
+	types.SetFolderUserSetting(conf, realFolder, types.SettingSnykIacEnabled, true)
 
 	resolver := testutil.DefaultConfigResolver(engine)
-	require.True(t, resolver.IsProductEnabledForFolder(product.ProductOpenSource, &types.FolderConfig{FolderPath: realFolder}))
-	require.False(t, resolver.IsProductEnabledForFolder(product.ProductOpenSource, &types.FolderConfig{FolderPath: referenceFolder}))
+	require.True(t, resolver.IsProductEnabledForFolder(product.ProductInfrastructureAsCode, &types.FolderConfig{FolderPath: realFolder}))
+	require.False(t, resolver.IsProductEnabledForFolder(product.ProductInfrastructureAsCode, &types.FolderConfig{FolderPath: referenceFolder}))
 
 	ctrl := gomock.NewController(t)
 	executor := mock_cli.NewMockExecutor(ctrl)
@@ -69,57 +69,57 @@ func newReferenceOSSScanner(
 		Execute(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 		DoAndReturn(func(context.Context, []string, types.FilePath, gotenv.Env) ([]byte, error) {
 			executed = true
-			return []byte("{}"), nil
+			return []byte("[]"), nil
 		}).
 		AnyTimes()
-	scanner := NewCLIScanner(
-		engine,
+	scanner := New(
+		conf,
+		engine.GetLogger(),
 		performance.NewInstrumentor(),
 		error_reporting.NewTestErrorReporter(engine),
 		executor,
-		getLearnMock(t),
-		notification.NewMockNotifier(),
 		resolver,
 		testutil.NewDrainedProgressTracker(),
-	).(*CLIScanner)
-	return scanner, &executed, resolver, referenceFolder
+	)
+	return scanner, &executed, resolver, referenceFolder, supportedPath
 }
 
-func TestCLIScanner_Scan_ReferenceBypassesOnlySyntheticFolderEnablement(t *testing.T) {
+func TestScanner_Scan_ReferenceBypassesOnlyTemporaryFolderEnablement(t *testing.T) {
 	tests := []struct {
 		name        string
-		scanTypes   []ctx2.DeltaScanType
+		options     []testutil.FolderScanContextOption
 		expectedErr string
 		executed    bool
 	}{
-		{name: "reference reaches CLI executor", scanTypes: []ctx2.DeltaScanType{ctx2.Reference}, executed: true},
-		{name: "working directory remains disabled", scanTypes: []ctx2.DeltaScanType{ctx2.WorkingDirectory}, expectedErr: utils.ErrSnykOssNotEnabledForFolder},
-		{name: "missing scan type remains disabled", expectedErr: utils.ErrSnykOssNotEnabledForFolder},
+		{name: "reference context overrules disabled and reaches CLI executor", options: []testutil.FolderScanContextOption{testutil.WithScanType(ctx2.Reference)}, executed: true},
+		{name: "working directory context and disabled returns not enabled error", options: []testutil.FolderScanContextOption{testutil.WithScanType(ctx2.WorkingDirectory)}, expectedErr: utils.ErrSnykIacNotEnabledForFolder},
+		{name: "missing scan type and disabled returns not enabled error", expectedErr: utils.ErrSnykIacNotEnabledForFolder},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			scanner, executed, resolver, referenceFolder := newReferenceOSSScanner(t, "valid-token")
+			scanner, executed, resolver, referenceFolder, supportedPath := newReferenceIaCScanner(t, "valid-token")
 
-			issues, err := scanner.Scan(testutil.ContextWithFolderScan(t, resolver, referenceFolder, tt.scanTypes...), referenceFolder)
+			issues, err := scanner.Scan(testutil.ContextWithFolderScan(t, resolver, referenceFolder, tt.options...), supportedPath)
 
 			if tt.expectedErr != "" {
 				require.EqualError(t, err, tt.expectedErr)
 				assert.Nil(t, issues)
 			} else {
 				require.NoError(t, err)
+				assert.Empty(t, issues)
 			}
 			assert.Equal(t, tt.executed, *executed)
 		})
 	}
 }
 
-func TestCLIScanner_Scan_ReferenceStillRequiresAuthentication(t *testing.T) {
-	scanner, executed, resolver, referenceFolder := newReferenceOSSScanner(t, "")
+func TestScanner_Scan_ReferenceRequiresAuthentication(t *testing.T) {
+	scanner, executed, resolver, referenceFolder, supportedPath := newReferenceIaCScanner(t, "")
 
 	issues, err := scanner.Scan(
-		testutil.ContextWithFolderScan(t, resolver, referenceFolder, ctx2.Reference),
-		referenceFolder,
+		testutil.ContextWithFolderScan(t, resolver, referenceFolder, testutil.WithScanType(ctx2.Reference)),
+		supportedPath,
 	)
 
 	require.EqualError(t, err, utils.MsgNotAuthenticatedNoScan)
@@ -127,13 +127,13 @@ func TestCLIScanner_Scan_ReferenceStillRequiresAuthentication(t *testing.T) {
 	assert.False(t, *executed)
 }
 
-func TestCLIScanner_Scan_ReferenceStillSkipsUnsupportedPath(t *testing.T) {
-	scanner, executed, resolver, referenceFolder := newReferenceOSSScanner(t, "valid-token")
-	unsupportedPath := types.FilePath(filepath.Join(string(referenceFolder), "main.go"))
-	require.NoError(t, os.WriteFile(string(unsupportedPath), []byte("package main"), 0o600))
+func TestScanner_Scan_ReferenceSkipsUnsupportedPath(t *testing.T) {
+	scanner, executed, resolver, referenceFolder, _ := newReferenceIaCScanner(t, "valid-token")
+	unsupportedPath := types.FilePath(filepath.Join(string(referenceFolder), "main.txt"))
+	require.NoError(t, os.WriteFile(string(unsupportedPath), []byte("not iac"), 0o600))
 
 	issues, err := scanner.Scan(
-		testutil.ContextWithFolderScan(t, resolver, referenceFolder, ctx2.Reference),
+		testutil.ContextWithFolderScan(t, resolver, referenceFolder, testutil.WithScanType(ctx2.Reference)),
 		unsupportedPath,
 	)
